+++ apps/logger_backend/lib/logger_backend.ex
@@ -1,0 +1,18 @@
+defmodule LoggerBackend do
+  @moduledoc """
+  Documentation for LoggerBackend.
+  """
+
+  @doc """
+  Hello world.
+
+  ## Examples
+
+      iex> LoggerBackend.hello()
+      :world
+
+  """
+  def hello do
+    :world
+  end
+end