+++ apps/logger_backend/test/logger_backend_test.exs
@@ -1,0 +1,8 @@
+defmodule LoggerBackendTest do
+  use ExUnit.Case
+  doctest LoggerBackend
+
+  test "greets the world" do
+    assert LoggerBackend.hello() == :world
+  end
+end